made certain lombok annotations assigned to annotationProces sor are added to classpath see build gradle file. Changed packge structure com.jasb... in entities and corresponding inports in classes of the other projects
