+++ toiletuserservice/src/main/java/com/toiletuserservice/repo/ToiletUserRepo.java
@@ -1,6 +1,6 @@
 package com.toiletuserservice.repo;
 
-import entities.ToiletUser;
+import com.jasb.entities.ToiletUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ToiletUserRepo extends JpaRepository<ToiletUser, Long> {
